update LineChart
change js line chart to apex line chart
delete linechart.js
